+++ src/components/Section/Section.jsx
@@ -4,19 +4,27 @@ import Header from '../Header';
 
 class Section extends Component {
   render() {
-    const { gray, title, subTitle, children, className, contentClassName } = this.props;
+    const {
+      gray,
+      title,
+      subTitle,
+      children,
+      className,
+      contentClassName,
+      alignHeader
+    } = this.props;
     return (
       <section className={`Section ${gray ? 'Section--gray' : ''} ${className}`}>
         <div className={`p-5 ${contentClassName}`}>
           {(title || subTitle) && (
             <div className="mb-5 text-center">
               {title && (
-                <Header textAlign="center" type="h2" className="Section__Title mb-2">
+                <Header textAlign={alignHeader} type="h2" className="Section__Title mb-2">
                   {title}
                 </Header>
               )}
               {subTitle && (
-                <Header textAlign="center" type="h3" className="Section__Subtitle mb-2">
+                <Header textAlign={alignHeader} type="h3" className="Section__Subtitle mb-2">
                   {subTitle}
                 </Header>
               )}
@@ -34,6 +42,7 @@ Section.propTypes = {
   title: PropTypes.string,
   subTitle: PropTypes.string,
   className: PropTypes.string,
+  alignHeader: PropTypes.string,
   contentClassName: PropTypes.string,
   children: PropTypes.oneOfType([PropTypes.arrayOf(PropTypes.node), PropTypes.node])
 };
@@ -42,6 +51,7 @@ Section.defaultProps = {
   title: '',
   subTitle: '',
   className: '',
+  alignHeader: 'left',
   contentClassName: '',
   children: null
 };