+++ storybook-utils/components/GridComponents/GridComponent.jsx
@@ -26,7 +26,7 @@ class GridComponent extends React.Component {
       <Row>
         {components.map(component => (
           <Column key={component} className="col-12 col-sm-6 col-md-4 col-lg-3">
-            <Link href={`/components/${component}`} tooltip="View more details">
+            <Link href={`/components/${component.toLowerCase()}`} tooltip="View more details">
               <Card
                 showBorder={false}
                 imageCircle={false}
@@ -59,7 +59,12 @@ class GridComponent extends React.Component {
     return (
       <Row className="p-3">
         {items.map(item => (
-          <Section key={item.name} title={item.name} subTitle={item.description}>
+          <Section
+            key={item.name}
+            className="grid-component"
+            title={item.name}
+            subTitle={item.description}
+          >
             {GridComponent.paintItems(item.components)}
           </Section>
         ))}