+++ changelog.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 #### Added
 
+- [#](https://github.com/front10/landing-page-book/pull/151) Added props `alignHeader` to  Section component, by [@jlfalcon3589]
 - [#151](https://github.com/front10/landing-page-book/pull/151) Props `circle`, `rounded`, `active`, `size`, `color`, `style` to component Button, by [@rootlinux2](https://github.com/rootlinux2)
 - [#151](https://github.com/front10/landing-page-book/pull/151) Props `btnRounded`, `btnDisabled` to component GithubButton, by [@rootlinux2](https://github.com/rootlinux2)
 - [#151](https://github.com/front10/landing-page-book/pull/151) CSS variable --GithubDetail_speech-rounded to component GithubButton, by [@rootlinux2](https://github.com/rootlinux2)